added functionality for forgto password, update password/user details

+++ controllers/courses.js
@@ -41,7 +41,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 exports.addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
     req.body.user = req.user.id;
-    const bootcamp = Bootcamp.findById(req.params.bootcampId);
+    const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if(!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with id: ${req.params.bootcampId}`, 404));
@@ -62,17 +62,26 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 // @route: PUT /api/v1/courses/:id
 // @access: Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
+    req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
+    
+
     let course = await Course.findById(req.params.id);
     if(!course) {
         return next(new ErrorResponse(`No course with id: ${req.params.id}`, 404));
     }
 
-     course = await Course.findByIdAndUpdate(req.params.id, req.body).populate({
+    // Make sure it is authorised owner
+    if(course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorised to update a course to ${course._id}`, 401)
+        )
+    }
+
+    course = await Course.findByIdAndUpdate(req.params.id, req.body).populate({
         path: "bootcamp",
         select: "name description"
     });
-
-    console.log("course: ", course)
 
     res.status(200).json({success: true, data: course})
 });
@@ -86,6 +95,14 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     if(!course) {
         return next(new ErrorResponse(`No course with id: ${req.params.id}`, 404));
     }
+
+    // Make sure it is authorised owner
+    if(course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorised to delete a course to ${course._id}`, 401)
+        )
+    }
+    
     await Course.findByIdAndDelete(req.params.id);
 
     res.status(200).json({success: true})
